Fix duplicate collision issue
We needed to set a delay on destroying blocks when hit by the ball, because the ball was not processing the collision before the block is destroyed. This introduced another bug where the block was counting as being destroyed twice.

This has been resolved by immediately removing the block from the blocks group, and adding it to the walls group (which does not get destroyed by collisions)

The initial collision when the block was part of the blockGroup is still triggered destroying the block after the delay, but until the delay happens its physics body stays around long enough for the ball to process collision with correctly.